completed basic HMAC auth between send & server
application server now generates an api key and secret for every user
this is display on login, on the new page views/dash.erb
requests to post or harvest messages now require an authenticate call,
which recalculates HMAC key and compares it to that which is now
sent by send_client.

also created a "key library" file, which handles reading and writing
api keys and secrets to a local file, but you still have to copy paste
them from the website.

updated model to handle api_key and secret, changed dev database name
for new environment.